+++ mobile/lib/widgets/asset_viewer/detail_panel/exif_map.dart
