+++ mobile/lib/widgets/asset_viewer/detail_panel/file_info.dart
@@ -3,10 +3,10 @@ import 'package:immich_mobile/entities/asset.entity.dart';
 import 'package:immich_mobile/extensions/build_context_extensions.dart';
 import 'package:immich_mobile/utils/bytes_units.dart';
 
-class AssetNameAndSize extends StatelessWidget {
+class FileInfo extends StatelessWidget {
   final Asset asset;
 
-  const AssetNameAndSize({
+  const FileInfo({
     super.key,
     required this.asset,
   });
