+++ mobile/lib/widgets/asset_viewer/detail_panel/people_info.dart
